requirements.txt
selenium>=4.17.2
webdriver-manager>=4.0.1
pillow>=10.3.0 